ovsdb: Fix leader test case failure on little CPU
In previous "reconnect to the leader" unit test case, there was no
mechanism to check whether the ovsdb servers have joined to the cluster
successfully, which led to the unit test failure on platforms with
little CPU. The root cause was not the failure of voting a new leader
after the original leader was killed but due to unsuccessful making
up a cluster at very beginning. This commit fixes this false-negative
testing failure.

Signed-off-by: Lance.Yang@arm.com <Lance.Yang@arm.com>